requirements.txt
django>=1.9.5
django-crispy-forms>=1.6.0
django-enumfield==1.3b2
